Wikipedia Data Extractor/requirements.txt
wikipedia-api
python-dotenv